Change automatic SDK installation folder
During the automatic SDK installation it was saved to
QStandardPaths::writableLocation(QStandardPaths::DataLocation), usually inside
~/.local/share/kdevelop/, inside a folder called arduino-ARDUINO_SDK_VERSION_NAME
(which is a constraint).
The problem begins once the plugin get an update and ARDUINO_SDK_VERSION_NAME changes
(which was already happened, 1.6.9 to 1.6.10). As we are searching for arduino inside
arduino-ARDUINO_SDK_VERSION_NAME, the previously downloaded SDK will not be visible
anymore.With this path, SDK will be saved to
QStandardPaths::writableLocation(QStandardPaths::DataLocation)/arduino so it won't get
lost anymore.

Please note, this change does not affect the internal SDK versioning, which is handled in
an another place and based on arduino/revisions.txt content, not on the folder name.

Signed-off-by: Alexander Zhigalin <alexander@zhigalin.tk>